IP Address functionality for multiplayer:
To play across the internet (2 computers):

Uncomment line 741 of GameApplication.java and simply run the server and
1 game application, then have your opponent edit your ip address (found
by running IPAddress.java) into line 742 of their GameApplication.java
and then run it.

For playing locally on a single computer:

Have line 742 of GameApplication uncommented (and 741 commented out) and
make the ip address equal to 127.0.0.1 (this is your localhost).  Just
run the server and then run two game applications.